Add q_reverseK function
Split the queue to 'q_size/k' segments first, then use
the 'seg_head' pointer to record the segment head address.

When having the segment head address, use list_add() to
add every element to the head of the segment. The effect
is equal to reverse the whole segment elements.

Change-Id: I725058dc3ada9ee910f97b7f40c33c1dc986c98d